refactor text_object into multiple files
