+++ hac-core/src/text_object/character.rs
@@ -1,0 +1,22 @@
+#[derive(PartialEq, Debug)]
+pub enum Kind {
+    Word,
+    Whitespace,
+    Punctuation,
+}
+
+pub fn kind(c: char) -> Kind {
+    match c {
+        _ if c.is_alphanumeric() => Kind::Word,
+        _ if c.is_whitespace() => Kind::Whitespace,
+        _ => Kind::Punctuation,
+    }
+}
+
+pub fn is_opening_token(char: char) -> bool {
+    matches!(char, '(' | '{' | '[' | '<')
+}
+
+pub fn is_closing_token(char: char) -> bool {
+    matches!(char, ')' | '}' | ']' | '>')
+}
